arm: mach-s5pv210: fix bigmem to correctly use memsizes
After the removal of fimc1, media dev 4 is jpeg.
So 12MiB was being allocated to jpeg in normal mode.
And 12MiB was still being reserved by fimc2 in nocam mode.

This changes ids from indexes to use enums to avoid confusion.
And also adds redunant, #defines for clarity.
Let's make it easier for possible future changes.

Requires, revert of bad6167010d928db4c07d514e98a1810c0269255
As that will cause overlapping and JPEG only takes 1MiB.
Anyways, there is no need for that after this fix.

However, we need to think about userspace,
As we require a hardcoded patch for OMX libs.
Which means even in normal mode, whole MFC is not mmaped
That might cause frame drops in hw/decoding and 720p recording.